[CFX-6668] address PR review comments
- rename internal/tls files to follow Go convention:
  windows.go → tls_windows.go, windows_stub.go → tls_other.go,
  windows_stub_test.go → tls_other_test.go
- fix PowerShell script: remove bare try/catch, rely on -ErrorAction SilentlyContinue
- extract TLS logic from cmd/root.go into cmd/tls.go
- split Windows-specific code into cmd/tls_windows.go (platform build tag):
  --export-windows-certs flag registration, windowsCACertPath (guards empty APPDATA),
  applyWindowsCerts
- cmd/tls_other.go: no-ops on non-Windows so flag and APPDATA logic never compile in
