+++ cmd/tls.go
@@ -1,0 +1,38 @@
+// Copyright 2026 DataRobot, Inc. and its affiliates.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package cmd
+
+import (
+	"github.com/datarobot/cli/internal/config/viperx"
+	internaltls "github.com/datarobot/cli/internal/tls"
+	"github.com/spf13/cobra"
+)
+
+// setupTLS configures http.DefaultTransport based on TLS-related flags and
+// the persisted ca-cert config value. Must run after initializeConfig so that
+// the ca-cert value from drconfig.yaml is available via viper.
+func setupTLS(cmd *cobra.Command) error {
+	skipVerify, _ := cmd.Flags().GetBool("skip-certificate-check")
+	caCert := viperx.GetString("ca-cert")
+
+	if err := applyWindowsCerts(cmd, &caCert); err != nil {
+		return err
+	}
+
+	return internaltls.Apply(internaltls.Options{
+		SkipVerify: skipVerify,
+		CACertPath: caCert,
+	})
+}
