+++ cmd/tls_other.go
@@ -1,0 +1,27 @@
+// Copyright 2026 DataRobot, Inc. and its affiliates.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:build !windows
+
+package cmd
+
+import "github.com/spf13/cobra"
+
+// registerExportWindowsCertsFlag is a no-op on non-Windows platforms.
+func registerExportWindowsCertsFlag(_ *cobra.Command) {}
+
+// applyWindowsCerts is a no-op on non-Windows platforms.
+func applyWindowsCerts(_ *cobra.Command, _ *string) error {
+	return nil
+}
