+++ internal/tls/tls_windows.go
@@ -30,13 +30,11 @@ func ExportWindowsCerts(dest string) error {
 		`$out = @()`,
 		`foreach ($store in 'Root','CA') {`,
 		`  foreach ($loc in 'LocalMachine','CurrentUser') {`,
-		`    try {`,
-		`      Get-ChildItem -Path "Cert:\$loc\$store" -ErrorAction SilentlyContinue | ForEach-Object {`,
-		`        $out += '-----BEGIN CERTIFICATE-----'`,
-		`        $out += [Convert]::ToBase64String($_.RawData, 'InsertLineBreaks')`,
-		`        $out += '-----END CERTIFICATE-----'`,
-		`      }`,
-		`    } catch {}`,
+		`    Get-ChildItem -Path "Cert:\$loc\$store" -ErrorAction SilentlyContinue | ForEach-Object {`,
+		`      $out += '-----BEGIN CERTIFICATE-----'`,
+		`      $out += [Convert]::ToBase64String($_.RawData, 'InsertLineBreaks')`,
+		`      $out += '-----END CERTIFICATE-----'`,
+		`    }`,
 		`  }`,
 		`}`,
 		`$out -join [Environment]::NewLine`,
